+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unpublished
+
+* (troessner) Remove `define_state_query_method` from public API
+* (troessner) Do not override existing methods when defining state query methods but warn the user.
+
 # 0.0.17 (2012-05-02):
 
 * (zmillman) Add write_state_without_persistence.